pyproject.toml
[project]
name = "discord-extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "discord>=2.3.2",
    "discord-py>=2.6.4",
    "lxml>=6.0.2",
    "markdownify>=1.2.2",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "urlextract>=1.9.0",
    "yt-dlp>=2026.2.21",
]
